Refactor service components and add debug code
