+++ frontend/src/pages/servicesComponents/DebugCode/index.js
@@ -1,0 +1,2 @@
+export * from "./DebugCode";
+export { default } from "./DebugCode";
