+++ frontend/src/pages/servicesComponents/ServicesComponents.jsx
@@ -1,67 +1,26 @@
-import React from "react";
 import App from "../../App";
 import ServiceReqTypeSelect from "./ServiceReqTypeSelect";
 import ServicesComponentForm from "./ServicesComponentForm";
 import ServicesComponentsList from "./ServicesComponentsList";
 import ConfirmationModal from "../../components/UI/Modals/ConfirmationModal";
+import { initialFormData } from "./servicesComponents.constants";
+import useServicesComponents from "./servicesComponents.hooks";
 
 const ServicesComponents = () => {
-    const [serviceTypeReq, setServiceTypeReq] = React.useState("");
-    const [formData, setFormData] = React.useState(initialFormData);
-    const [servicesComponents, setServicesComponents] = React.useState([initialServicesComponent]);
-    const [showModal, setShowModal] = React.useState(false);
-    const [modalProps, setModalProps] = React.useState({});
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        if (formData.mode === "add") {
-            const newFormData = {
-                id: crypto.randomUUID(),
-                ...formData
-            };
-            setServicesComponents([...servicesComponents, newFormData]);
-            alert("Form submitted");
-            setFormData(initialFormData);
-        }
-        if (formData.mode === "edit") {
-            handleEdit(formData.id);
-            alert("Services Component updated");
-            setFormData(initialFormData);
-        }
-    };
-
-    const handleEdit = (id) => {
-        const index = servicesComponents.findIndex((component) => component.id === id);
-        const newServicesComponents = [...servicesComponents];
-        const newFormData = { ...formData, mode: "add" };
-        newServicesComponents[index] = { ...servicesComponents[index], ...formData };
-        setServicesComponents(newServicesComponents);
-        setFormData(newFormData);
-    };
-
-    const handleDelete = (id) => {
-        const newServicesComponents = servicesComponents.filter((component) => component.id !== id);
-        setShowModal(true);
-        setModalProps({
-            heading: "Are you sure you want to delete this Services Component?",
-            actionButtonText: "Delete",
-            secondaryButtonText: "Cancel",
-            handleConfirm: () => {
-                setServicesComponents(newServicesComponents);
-            }
-        });
-    };
-
-    const handleCancel = (e) => {
-        e.preventDefault();
-        setFormData(initialFormData);
-    };
-
-    const setFormDataById = (id) => {
-        const index = servicesComponents.findIndex((component) => component.id === id);
-        const newFormData = { ...servicesComponents[index], mode: "edit" };
-        setFormData(newFormData);
-    };
+    const {
+        formData,
+        modalProps,
+        serviceTypeReq,
+        servicesComponents,
+        setFormData,
+        setServiceTypeReq,
+        setShowModal,
+        showModal,
+        handleSubmit,
+        handleDelete,
+        handleCancel,
+        setFormDataById
+    } = useServicesComponents();
 
     return (
         <App breadCrumbName="Playground">
@@ -90,12 +49,6 @@ const ServicesComponents = () => {
                     handleSubmit={handleSubmit}
                     handleCancel={handleCancel}
                 />
-                {import.meta.env.DEV && (
-                    <section className="border-dashed border-emergency margin-top-6">
-                        <h2 className="margin-0">Form Data</h2>
-                        <pre>{JSON.stringify(formData, null, 2)}</pre>
-                    </section>
-                )}
             </section>
             <ServicesComponentsList
                 servicesComponents={servicesComponents}
@@ -104,33 +57,6 @@ const ServicesComponents = () => {
             />
         </App>
     );
-};
-
-const initialServicesComponent = {
-    id: "9ab46509-8a3c-498a-998d-b40e78df5cd3",
-    servicesComponent: "SC1",
-    optional: false,
-    popStartMonth: "3",
-    popStartDay: "15",
-    popStartYear: "2024",
-    popEndMonth: "1",
-    popEndDay: "15",
-    popEndYear: "2025",
-    description:
-        "Develop a theory of change and identify ways to improve the program through continuous user feedback and engagement"
-};
-
-const initialFormData = {
-    servicesComponent: "",
-    optional: "",
-    popStartMonth: "",
-    popStartDay: "",
-    popStartYear: "",
-    popEndMonth: "",
-    popEndDay: "",
-    popEndYear: "",
-    description: "",
-    mode: "add"
 };
 
 export default ServicesComponents;
