+++ frontend/src/pages/servicesComponents/ServicesComponentsList/ServicesComponentsList.jsx
@@ -1,4 +1,5 @@
 import ServicesComponentListItem from "../ServicesComponentListItem";
+import DebugCode from "../DebugCode";
 
 function ServicesComponentsList({ servicesComponents, setFormDataById, handleDelete }) {
     return (
@@ -15,12 +16,10 @@ function ServicesComponentsList({ servicesComponents, setFormDataById, handleDel
             ) : (
                 <p>You have not added any Services Component yet.</p>
             )}
-            {import.meta.env.DEV && (
-                <section className="border-dashed border-emergency margin-top-6">
-                    <h2 className="margin-0">Services Components</h2>
-                    <pre>{JSON.stringify(servicesComponents, null, 2)}</pre>
-                </section>
-            )}
+            <DebugCode
+                title="Services Components"
+                data={servicesComponents}
+            />
         </section>
     );
 }
