+++ frontend/src/pages/servicesComponents/servicesComponents.constants.js
@@ -1,0 +1,36 @@
+export const initialServicesComponent = {
+    id: "9ab46509-8a3c-498a-998d-b40e78df5cd3",
+    servicesComponent: "SC1",
+    optional: false,
+    popStartMonth: "3",
+    popStartDay: "15",
+    popStartYear: "2024",
+    popEndMonth: "1",
+    popEndDay: "15",
+    popEndYear: "2025",
+    description:
+        "Develop a theory of change and identify ways to improve the program through continuous user feedback and engagement"
+};
+
+export const initialFormData = {
+    servicesComponent: "",
+    optional: "",
+    popStartMonth: "",
+    popStartDay: "",
+    popStartYear: "",
+    popEndMonth: "",
+    popEndDay: "",
+    popEndYear: "",
+    description: "",
+    mode: "add"
+};
+
+export const nonSeverableOptions = ["SC1", "SC2", "SC3", "SC4", "SC5", "SC6"];
+export const severableOptions = [
+    "Base Period 1",
+    "Option Period 2",
+    "Option Period 3",
+    "Option Period 4",
+    "Option Period 5",
+    "Option Period 6"
+];
