+++ frontend/src/pages/servicesComponents/servicesComponents.helpers.js
@@ -1,0 +1,3 @@
+export const addAnOInFront = (str) => {
+    return "O" + str;
+};
